rename bzt.modules.selenium
